Extend lib/python, add 2 new tests.
In lib/python, add the ability to create multiple running instances of Asterisk
from within a python script.  This was inspired by the same feature provided by
asttest for tests written in lua.

Add the 'ami-login' test.  This uses Asterisk() from lib/python/ to create an
instance of Asterisk and ensure that it can log in to the AMI interface
successfully.

Add the 'iax-call-basic' test.  This creates 2 Asterisk() instances and originates
an IAX2 call between them.  Both ends connect to a FastAGI server implemented by
the test script to verify that the call was set up.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@162 ad95026d-47f8-4470-b445-30a6b62014f3