Add "sendfullybooted" option to default asterisk.conf file.
Without this, tests that run on non-trunk branches will not hang forever
due to their inability to catch the FullyBooted event.

A further modification would be to make use of the CLI command in the
asterisk:spawn() function so that it will not return until fully booted.
Then there would be no need to have individual tests use an AMI connection
for that purpose.



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@353 ad95026d-47f8-4470-b445-30a6b62014f3