Change the sleep time for queue-baseline test.
The automated tests were failing sometimes due to not receiving
the AgentComplete event. This event is sent at the end of the call. To
ensure that this event is received, we should sleep for longer than the
duration of the call just to be on the safe side. The sleep is now
1 second longer than the call's duration.

Also added a bit of debugging so that if this test fails again, we
can see more clearly why.



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@363 ad95026d-47f8-4470-b445-30a6b62014f3