Added leave_voicemail_nominal test; support for test failure as an expected result
Adds an initial test for app_voicemail, specifically the nominal execution of the VoiceMail
dialplan application.  Added support in test.yaml for properties/expectedResult option, which
will accept a value of 'Fail' to indicate that the test is expected to fail.  If the test fails,
a failure is not reported to the test execution engine and the test passes; otherwise, if the
test passes, a failure is reported to the test engine.

Review: https://reviewboard.asterisk.org/r/1354/



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@1833 ad95026d-47f8-4470-b445-30a6b62014f3