Remove OptionParser from tests
This patch removes the ability to pass command-line arguements to test. Not only
is this a sloppy way of doing things, 95% of the time is was not even required.
Because some tests used the ast_version from the argv parameter, we've added the
cli_aliases.conf file to help with CLI syntax.

Additionall, I've removed valgrid support (was not used for anything) until we
decided how to find a better solution for it.  Ideally, the testsuite will have
a .config file all tests can access for global settings.

Review: https://reviewboard.asterisk.org/r/1386/


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@1962 ad95026d-47f8-4470-b445-30a6b62014f3