Patch to fix bouncing test check_voicemail_new_user
This patch to the voicemail test base class will hopefully fix the bouncing test
check_voicemail_new_user.  When performing the new_user operations in voicemail
where the sound greetings are being recorded, there are some fairly tight timing
constraints in which different DTMF tones must be sent.  Because these were being
set along with the TALK_AUDIO variable, at times the user would timeout and be
prematurely kicked out of voicemail.  This patch caches the last set TALK_AUDIO and
DTMF_TO_SEND variables to reduce the number of calls to cli_exec.  If this doesn't
fix the timing issue, a special extension can be written for that test to bypass
the useage of variables.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@2280 ad95026d-47f8-4470-b445-30a6b62014f3