Adds tests for 'queue set penalty' command and analogous AMI command
This test basically just has a few preset queue members in a number of queues and the
test is set up to test different uses of the command as a state machine. It then checks
QueuePenalty events sent in response to the commands and makes sure they are in line with
expectations. This test was made in response to a regression that was introduced when
adding realtime support to queue set penalty.

Review: https://reviewboard.asterisk.org/r/1611/


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@2907 ad95026d-47f8-4470-b445-30a6b62014f3