Change Asterisk start to look for return value instead of exception
Asterisk start depended on cli_exec throwing an exception to know
when to retry.  This caused problems when cli_exec would throw an
exception elsewhere in the testsuite, as those parts of the code
that were inside a twisted reactor callback could cause an exception
in a portion of those code that wasn't handled, killing the reactor
but not propagating to the top-most level.  The Asterisk start now
checks the return value - if no output is returned, it tries again.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@2983 ad95026d-47f8-4470-b445-30a6b62014f3