Perform initial tagging of tests
This patch builds on r2724 by adding tags to the vast majority of tests.
In addition, it adds a new command line parameter, '-L', which lists
the available tags in the Test Suite.  A tag is counted as being 'available'
if there exists at least one test that will be executed if that tag parameter
is the only tag parameter passed to runtests.

Review: https://reviewboard.asterisk.org/r/1791/



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3077 ad95026d-47f8-4470-b445-30a6b62014f3