Migrate two fastagi tests to use twisted for process management
Tests fastagi/execute and fastagi/wait-for-digit were missed in the
migration patch, due to the fact that neither actually failed.  In
the case of wait-for-digit, that's because its not executed; execute
did not fail despite the fact that Asterisk never started.  This patch
migrates them to use twisted for their process management in the same
manner as the other non-TestCase derived fastagi tests.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3135 ad95026d-47f8-4470-b445-30a6b62014f3