Make ami_login_error return a twisted Failure
Per the specification for deferds, an errback must either raise
an exception or return a twisted.python.failure.Failure instance.
Otherwise, the usual chain of callbacks is returned to.  This isn't
terribly useful when the error in question is a failure to connect
to AMI.  This should make the login error a bit more terminal when
something goes wrong during an AMI connection.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3152 ad95026d-47f8-4470-b445-30a6b62014f3