Update the pre-/post-condition check framework to support twisted
This patch updates the pre-/post-condition checking framework in the
Asterisk Test Suite to use the twisted asynchronous framework for its
CLI/AMI commands.  This was necessary, as the starting/stopping
of Asterisk now expects things that explicitly need to be involved
in the process to be in the deferred callback chain.

Review: https://reviewboard.asterisk.org/r/1831

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3188 ad95026d-47f8-4470-b445-30a6b62014f3