Add support for Certified Asterisk branches
Revision 3205 added support for Certified Asterisk tags, but not for
the actual branch.  Since the branch is based off of a particular tag
of Asterisk (e.g., 1.8.11), branch testing has to assume that the 
capabilities mirror that of Asterisk 1.8.11.X, as opposed to the main
1.8 branch.  This modifies the version parsing to treat SVN branches that
include a minor version (which all Certified Asterisk branches will have)
as a tag of that version for the purposes of min/max version checking.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3208 ad95026d-47f8-4470-b445-30a6b62014f3