Add ConfBridge Recording Test
This test checks for recording of a conference in ConfBridge via
the MixMonitor application.  It tests enabling and setting the location
of the recording via the CONFBRIDGE function and via the confbridge.conf
configuration file.  After a conference is recorded, the test verifies
that the recorded file contains some audio via the BackgroundDetect
application.

Review: https://reviewboard.asterisk.org/r/1880/

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3213 ad95026d-47f8-4470-b445-30a6b62014f3