Fix the sip_cause test
It appears that the dialplan post-Dial is executing faster than the
SIP_CAUSE hash can be set via the MASTER_CHANNEL() dialplan function.
This is causing the user event with cause information to be reported
before the final value is set into SIP_CAUSE. The Wait(2) (gross, I
know) should work around this problem.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3227 ad95026d-47f8-4470-b445-30a6b62014f3