Disable masquerade test until failures are debugged
It appears as if having the timing source be timer_fd is causing
local channel optimizations to hang.  This does not appear to be the
case with DAHDI as the timing source - so until this can be debugged
further, the test is being disabled.  ASTERISK-19966 created to track
and resolve this problem.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3260 ad95026d-47f8-4470-b445-30a6b62014f3