Fix core dump detection and force core files to fail tests
Previously, the core dump detection would only find core files if
they were explicitly named 'core'.  This doesn't work when the core
files are appended with the process ID.  Additionally, multiple core
dumps from a single test - which can occur with multiple instances of
Asterisk - would result in a single backtrace.

This patch looks for any file name 'core', regardless of the extension,
in the running directory and attempts to archive it.  The archiving now
trys to limit collisions in the backtraces.  Finally, at the end of each
test we check for core dumps and fail the test if any is found, as opposed
to the previous mechanism which would only look for core files if a test
failed.  Note that if the archiving process fails, all tests will end up
failing.  Make sure you run the Test Suite with sufficient permissions to do
its job. 

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3286 ad95026d-47f8-4470-b445-30a6b62014f3