Convert sip_attended_transfer test to use TestCase and PJSUA
This test has been disabled for a long time now because it was failing.
This change updates the test to use the TestCase class to bring it in line
with current practices and changes it over to using the pjsua python
library to avoid threading issues with the twisted reactor as well as
using a nicer interface to access the required features.

(closes issue SWP-4659)


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3346 ad95026d-47f8-4470-b445-30a6b62014f3