Work around pjsua socket release failure
When pjsua is used with TCP sockets, it fails to release the socket
upon exit instead causing the socket to be in use for up to 60 seconds
after pjsua exits before the socket times out. This change moves the
ports used for TCP sockets outside the range normally used for pjsua
and pjsip instances. This also re-enables the test.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3462 ad95026d-47f8-4470-b445-30a6b62014f3