Delay hanging up calls in bridge tests if the calls are not currently bridged.
This is expected to clear up some timing-related failures Jonathan Rose is
seeing while developing automon tests. The suspicion is that the hangup is
being issued while the bridge is broken and the automon is being stopped.
If we wait until the bridge is re-formed, then the hangup should be handled
properly.

The only reason this is required is due to two issues being seen in the test

1) Sometimes there is no BRIDGE_END CEL event issued.
2) When the hangup is issued while the bridge is broken, the CEL HANGUP event
   has incorrect hangupsource set.



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3580 ad95026d-47f8-4470-b445-30a6b62014f3