Testsuite: Call Parking timeouts and failure tokens
For the parkcall timeout tests, there are two main categories hinging on
whether comebacktoorigin is enabled or disabled for a given parkinglot.

In the case of comebacktoorigin=no tests, the primary things checked are
that a dial extension is created in the park-dial context and that the
call proceeds to a predictable position in the dialplan based on what
the timeout context is as well as what extensions exist within it. Also
a number of parking based variables that are supposed to be set on timeout
are checked to see if they are consistent with the expectations.

For comebacktoorigin=yes tests, each call will automatically dial the
extension to call back the parker in park-dial. This being the case, what
is checked here is that the channel continues through the PBX in a
consistent manner based on how the parker responds to the timed out call.
The three possibe cases are an answer followed by a hangup (nominal),
letting the dial timeout, and sending a busy indication. These tests
currently test the functionality of dial more so than call parking, but
they are in anticipation of possible changes to bridging that may not
necessarily involve directly invoking the dial application from the PBX.

As a side note, to support these tests I added a new feature to TestCase
called a fail token. These are useful for multiple module tests because
it is possible for one module to set success, and from that point on
unless another module explicitly sets failure, the test is successful.
That's less than ideal since the idea is normally that a module should
explicitly set success when it is successful and otherwise it is a failed
module, so failure tokens allow a module to communicate to a testcase that
unless it explicitly removes that token that the test should be considered
failed.  This is especially useful when a module throws an exception or
just doesn't reach a condition for establishing success/failure for
whatever reason.

Review: https://reviewboard.asterisk.org/r/2306/
Review: https://reviewboard.asterisk.org/r/2302/
Review: https://reviewboard.asterisk.org/r/2301/



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3633 ad95026d-47f8-4470-b445-30a6b62014f3