Refactored the call files test.
The call files test was not taking advantage of the testing framework, thus it
duplicated much of the functionality found already in the framework.  Refactored
the call files test to use the testing framework (e.g. it now inherits from
TestCase) and the standard python logging facilities.

Reported by: Matt Jordan



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3637 ad95026d-47f8-4470-b445-30a6b62014f3