Add bridging test that uses the blind transfer feature to park a channel.
This test uses BridgeTestCase to setup a bridge between two endpoints and
utilizes the DTMF blind transfer feature to ensure the feature is detected. It
also ensures that the transfered channels to the default, a static, and a
dynamic parking lot are successful when the lots are not full. Lastly it fills
a static and dynamic parking lot to ensure that further parking attempts fail.
The AMI callback type is used to hang up the channels that were successfully
and unsuccessfully parked.

Review: https://reviewboard.asterisk.org/r/2458/



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3766 ad95026d-47f8-4470-b445-30a6b62014f3