Add pluggable CEL verification module
This uses AMI to allow verification of CEL messages. It can be configured to
allow CEL messages to come before and/or after others if they are not always
in the same order due to channels operating on their own threads of execution.

(issue ASTERISK-21422)
(closes issue ASTERISK-21422)

Review: https://reviewboard.asterisk.org/r/2498/



git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3783 ad95026d-47f8-4470-b445-30a6b62014f3