Don't create a race condition between hanging up ast1 and ast2
In the SIP TLS test, Asterisk instance 1 sends a DTMF tone to Asterisk
instance 2 and waits for the second instance to break it out of a
Background application. Asterisk instance 2 sends a DTMF and then hangs
up. The problem is the hangup of Asterisk instance 2 may cause instance
1 to hangup prior to servicing the DTMF. This race condition always
existed; however, with the DTMF/AMI changes, it's much more likely to hit
it.

Now only the last instance calls Hangup. The other instance will Echo until
hung up.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3791 ad95026d-47f8-4470-b445-30a6b62014f3