Allow a SimpleTestCase object to specify a location for config files
This allows a group of tests to specify a shared common repository for
config files to use. This is of particular use for classes of tests that
need the same configuration. Note that this path is used in addition to
the normal location for configuration files, and supplements whatever is
in that location.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3853 ad95026d-47f8-4470-b445-30a6b62014f3