Add a hangup event instance class to PluggableModules
It's a fairly common need to arbitrarily hang up channel. This patch
adds two modules to do this. The first will hangup a channel given an
AMI event that contains the channel to hang up. The second is less picky -
it tracks all channels and, upon matching on some arbitrary event - hangs
up everyone.

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3860 ad95026d-47f8-4470-b445-30a6b62014f3