Allow the Asterisk Test Suite to run outside of an Asterisk source directory
For a host of historical reasons, the Asterisk Test Suite used to have to run as a
subdirectory inside of the Asterisk source.

This actually hasn't been needed for quite some time. The last hold out *was* the
buildoptions library; however, as it turns out, buildoptions will check the installed
Asterisk headers for the build options it needs to execute.

This patch removes the checks and updates the README. It also makes the buildoptions
library raise an Exception if it couldn't find the header - this really is needed, as
certain build options (such as TEST_FRAMEWORK) are needed for a large number of tests to
execute.

Review: https://reviewboard.asterisk.org/r/2659

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3900 ad95026d-47f8-4470-b445-30a6b62014f3