Update version string parsing to handle Debian style version suffixes
If an Asterisk version string follows the following nomenclature:
    "1.8.13.1~dfsg-2"
An exception will be thrown.

The version string parsing assumes that for "normal" Asterisk versions,
any additional tags added to a version are a "feature", and that those
features are denoted by a hyphen as opposed to a tilde. Debian will
append a "~dfsg[<num>]" when a source tarball needs to be repackaged,
normally due to not complying with the DFSG (thank you Tzafrir for
the explanation!)

This patch updates the version parsing to split the version string into
number and feature parts on both a '~' and a '-'. It also adds dfsg to
the accepted "features" list.

(closes issue ASTERISK-21320)
Reported by: Tzafrir Cohen
patches:
  Fix-parsing-a-dfsg-version.patch uploaded by Tzafrir Cohen (license 5035)

git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3913 ad95026d-47f8-4470-b445-30a6b62014f3