Let tests execute if any tag specified is matched
Originally, tags were used to further exclude tests. This meant that
specifying 'CDR' and 'SIP' would only execute a test if the test had both the
'CDR' and the 'SIP' flag. While this is useful to filter large sets of tests,
it is actually less useful in practice. Often, it is more useful to execute
different categories of tests together rather than execute a subset of an
existing category.

For example, it is useful to be able to execute all channel driver tests. This
lets you specify '-g SIP -g IAX (etc.)' on the command line and execute all of
the channel driver tests. Previously, this would only have executed a test if
it was both SIP and IAX (which is highly unlikely).


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@3961 ad95026d-47f8-4470-b445-30a6b62014f3