Account for ordering of channel destruction in expected results
The Local channels now (apparently!) have a slightly different ordering when
they nip off and get themselves hung up. As such, the CDR results were off.
The expected results now account for the order in which the channels will
be hung up based on how they are hung up and kicked out from bridge to bridge.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4036 ad95026d-47f8-4470-b445-30a6b62014f3