Turn CDR test originations into asynchronous originations
A synchronous origination will block all Stasis messages regarding that channel
until the call has finished. This has serious repercussions for the CDR tests -
when the channel (finally) hangs up, the dam of Stasis messages is released,
and both AMI and CDR recipients finally get the blocked up messages.
Unfortunately, AMI will often throw the Hangup event out while the CDR engine
is still frantically trying to catch up, resulting in Asterisk being shut
down just as it is attempting to finalize out the last CDR.

This wouldn't be a problem, as the engine will still attempt to spit out the
last message, but the CDR backends themselves will usually unload before that
happens. Sadness.

This patch configures the tests to Originate their calls asynchronously. In
addition to being the "right" thing to do, it also means that when a channel
hangs up, the CDR engine just has to process a single message, as opposed to
100.


git-svn-id: file:///home/opticron/private_mirror/testsuite_secondary/asterisk/trunk@4072 ad95026d-47f8-4470-b445-30a6b62014f3